change data type form long to double

+++ src/main/java/domain/calculator/Calculator.java
@@ -1,5 +1,5 @@
 package domain.calculator;
 
 public interface Calculator {
-    long calculate(long purchaseAmount);
+    double calculate(double purchaseAmount);
 }
