+++ src/test/java/acceptancetests/UserStory_1.java
@@ -10,6 +10,7 @@ import models.CustomerType;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -25,7 +26,7 @@ public class UserStory_1 {
 
     @Test
     public void given_discount_slabs_for_purchase_amount_5000_bill_amount_should_be_5000() {
-        long purchaseAmount = 5000;
+        double purchaseAmount = 5000;
         CartDiscountCalculator discountCalculator = new CartDiscountCalculator(discountSlabs);
 
         double billAmount = discountCalculator.billAmount(CustomerType.REGULAR, purchaseAmount);
@@ -35,7 +36,7 @@ public class UserStory_1 {
 
     @Test
     public void given_discount_slabs_for_purchase_amount_10000_bill_amount_should_be_9500() {
-        long purchaseAmount = 10000;
+        double purchaseAmount = 10000;
         CartDiscountCalculator discountCalculator = new CartDiscountCalculator(discountSlabs);
 
         double billAmount = discountCalculator.billAmount(CustomerType.REGULAR, purchaseAmount);
@@ -45,11 +46,11 @@ public class UserStory_1 {
 
     @Test
     public void given_discount_slabs_for_purchase_amount_15000_bill_amount_should_be_13500() {
-        long purchaseAmount = 15000;
+        double purchaseAmount = 15000;
         CartDiscountCalculator discountCalculator = new CartDiscountCalculator(discountSlabs);
 
         double billAmount = discountCalculator.billAmount(CustomerType.REGULAR, purchaseAmount);
 
-        Assert.assertEquals(13500, billAmount, 0);
+        Assert.assertEquals(13500, billAmount,0);
     }
 }
